pyproject.toml
[project]
name = "ai-influencer-agency"
version = "0.1.0"
description = "Interview Agent with Human-in-the-Loop"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.40",
    "langgraph-cli[inmem]>=0.3.1",
    "numpy>=2.3.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.0.0",
]
